move ZeroCopyCheckExporter to make bakeddata
