+++ tools/make/bakeddata/Cargo.toml
@@ -10,13 +10,15 @@ edition.workspace = true
 
 [dependencies]
 icu = { workspace = true, features = ["unstable"] }
-icu_provider = { workspace = true }
+icu_provider = { workspace = true, features = ["deserialize_postcard_1"] }
 icu_provider_export = { workspace = true, features = ["baked_exporter", "rayon"] }
 icu_provider_source = { workspace = true, features = ["networking", "unstable", "use_wasm"] }
+icu_provider_registry = { workspace = true }
 
 log = { workspace = true }
 simple_logger = { workspace = true }
 crlify = { workspace = true }
+postcard = { workspace = true, features = ["alloc"] }
 
 [package.metadata.cargo-all-features]
 # Unpublished internal tool; skip
